src/app5/requirements.txt
# Core requirements for app5.py
python-dotenv>=1.0.0
azure-identity>=1.15.0
agent-framework-azure-ai --pre
# Note: The --pre flag is required while Agent Framework is in preview

# Observability requirements
azure-ai-projects
azure-ai-agents